// frontend/src/lib/generated/types.gen.ts
export type ElementsLabelEntry = {
  /**
   * A string field
   */
  icon?: string;
  iconPosition: IconPositionEnum;
  /**
   * A text field
   */
  text: string;
  /**
   * A string field
   */
  ariaDescription: string;
  /**
   * Component instance ID
   */
  id?: number;
};

export type ElementsHeaderEntry = {
  alignment: AlignmentEnum;
  /**
   * A boolean field
   */
  promoteHeaderIcon: boolean | null;
  /**
   * A component field
   */
  header?: ElementsLabelEntry;
  /**
   * A component field
   */
  subheader?: ElementsLabelEntry;
  /**
   * Component instance ID
   */
  id?: number;
};

export type ElementsButtonEntry = {
  /**
   * A string field
   */
  url: string;
  /**
   * A boolean field
   */
  openInNewTab: boolean | null;
  /**
   * A component field
   */
  label?: ElementsLabelEntry;
  /**
   * Component instance ID
   */
  id?: number;
};
// ...
export type PluginUploadFileDocument = {
  /**
   * The unique document identifier as a UUID (v1-v8 or nil UUID). This ID persists across draft/published versions and localizations of the same document.
   */
  documentId: string;
  id: string | number;
  /**
   * A string field
   */
  name: string;
  /**
   * A text field
   */
  alternativeText?: string;
  /**
   * A text field
   */
  caption?: string;
  /**
   * A JSON field
   */
  focalPoint?: unknown;
  /**
   * An integer field
   */
  width?: number;
  /**
   * An integer field
   */
  height?: number;
  /**
   * A JSON field
   */
  formats?: unknown;
  /**
   * A string field
   */
  hash: string;
  /**
   * A string field
   */
  ext?: string;
  /**
   * A string field
   */
  mime: string;
  /**
   * A decimal field
   */
  size: number;
  /**
   * A text field
   */
  url: string;
  /**
   * A text field
   */
  previewUrl?: string;
  /**
   * A string field
   */
  provider: string;
  /**
   * A JSON field
   */
  provider_metadata?: unknown;
  /**
   * Timestamp when this entry was first created in the CMS.
   */
  createdAt?: string;
  /**
   * Timestamp when this entry was last modified.
   */
  updatedAt?: string;
  /**
   * Timestamp when this entry was published. Null for draft entries. Part of Strapi Draft & Publish feature.
   */
  publishedAt: string | null;
  related?: unknown;
};
// ...
export type SectionsHeroEntry = {
  /**
   * Component instance ID
   */
  id: number;
  /**
   * A component field
   */
  header: ElementsHeaderEntry;
  /**
   * A component field
   */
  exploreButton?: ElementsButtonEntry;
  /**
   * A media field
   */
  image: PluginUploadFileDocument;
  variant: VariantEnum;
};
// ...
export enum DirectionEnum {
  LTR = "ltr",
  RTL = "rtl",
}

/**
 * An enum field
 */
export enum IconPositionEnum {
  BEFORE_TEXT = "before_text",
  AFTER_TEXT = "after_text",
}

/**
 * An enum field
 */
export enum AlignmentEnum {
  CENTER = "center",
  LANGUAGE_DIRECTION = "language-direction",
}

/**
 * An enumeration field
 */
export enum VariantEnum {
  TEXT_ABOVE_BACKGROUND = "text-above-background",
  TEXT_OVER_BACKGROUND = "text-over-background",
  TEXT_BELOW_BACKGROUND = "text-below-background",
}